Extract test buffer utility
