+++ rust_extension/tests/test_utils/shared_buffer.rs
@@ -1,0 +1,19 @@
+use crate::{Arc, Mutex};
+use std::io::{self, Write};
+
+#[derive(Clone)]
+pub struct SharedBuf(pub Arc<Mutex<Vec<u8>>>);
+
+impl Write for SharedBuf {
+    fn write(&mut self, buf: &[u8]) -> io::Result<usize> {
+        self.0.lock().unwrap().write(buf)
+    }
+
+    fn flush(&mut self) -> io::Result<()> {
+        self.0.lock().unwrap().flush()
+    }
+}
+
+pub fn read_output(buffer: &Arc<Mutex<Vec<u8>>>) -> String {
+    String::from_utf8(buffer.lock().unwrap().clone()).unwrap()
+}
